Move `getBlockCategoriesAndSelecrors`
Move `getBlockCategoriesAndSelecrors`to styles extensions
